% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for R Shiny app and for making figures}
\usage{
generate_data(
  data.directory = data.directory,
  repository = repository,
  overwrite = FALSE
)
}
\arguments{
\item{data.directory}{Directory where the nGSL datasets are located.}

\item{repository}{Repository folder where files generated by the function are stored.}

\item{overwrite}{\code{TRUE} to overwrite existing files.}
}
\value{
(invisible) Files created and stored.
}
\description{
This function calculate the annual values for each indicator and generate the files or data needed to update the figures
from Isabel et al. (2025) or to update the figures in the Shiny app.
}
\examples{
# example code
generate_data(data.directory = "//dcqcimlna01a//BD_Peches//Releves_Poissons_de_Fond_et_Crevette//Donnees_PACES//", repository = "data/", overwrite = TRUE)

}
