% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_figures.R
\name{generate_figures}
\alias{generate_figures}
\title{Generate figures to update technical report indicators}
\usage{
generate_figures(repository, figures, overwrite = FALSE)
}
\arguments{
\item{repository}{Repository folder where files generated "generate_data" function are stored.}

\item{figures}{Repository folder where figures generated by the function are stored.}

\item{overwrite}{\code{TRUE} to overwrite existing files.}
}
\value{
(invisible) Figures created and stored.
}
\description{
This function generate the figures published in Isabel et al. (2025).
}
\examples{
# example code
generate_figures(repository = "data/", figures = "fig/", overwrite = TRUE)

}
